More resources, added logo to Nav, fixed bug in topic deletion in myInbox ('Pressing the delete button would delete all even if no items were selected)

+++ app/Http/Controllers/Frontend/myInboxController.php
@@ -223,11 +223,15 @@ class myInboxController extends Controller
     {
         $user = \Auth::user();
 
-        $user->threads()->detach($request->delete);
+        if(is_Null($request->delete))
+        {
+            \Notification::warning('No Conversations selected');
+        } else {
+            $user->threads()->detach($request->delete);
+            \Notification::success('You have left these conversations');
 
-        \Notification::success('You have left these conversations');
+        }
         return redirect('/my-inbox');
-
     }
 
 }
